pyproject.toml
[project]
name = "aiassistantbackend"
version = "0.1.0"
description = "Backend for AI Assistant"
requires-python = ">=3.12"
dependencies = [
    "authlib>=1.5.1",
    "fastapi[standard]>=0.115.11",
    "flask>=3.1.0",
    "flask-alembic>=3.1.1",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "pydantic-ai[logfire]>=0.0.25",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "sqlmodel>=0.0.23",
]
